Move DuckAddressStatusRepository to impl module
